move depretate module to bokeh.util
